Fix #373
