+++ crates/emmylua_code_analysis/src/db_index/type/type_ops/remove_type.rs
@@ -1,5 +1,7 @@
 use crate::{DbIndex, LuaType, LuaUnionType};
 
+use super::get_real_type;
+
 pub fn remove_type(db: &DbIndex, source: LuaType, removed_type: LuaType) -> Option<LuaType> {
     if source == removed_type {
         match source {
@@ -162,17 +164,4 @@ pub fn remove_type(db: &DbIndex, source: LuaType, removed_type: LuaType) -> Opti
     }
 
     Some(source)
-}
-
-fn get_real_type<'a>(db: &'a DbIndex, compact_type: &'a LuaType) -> Option<&'a LuaType> {
-    match compact_type {
-        LuaType::Ref(type_decl_id) => {
-            let type_decl = db.get_type_index().get_type_decl(type_decl_id)?;
-            if type_decl.is_alias() {
-                return get_real_type(db, type_decl.get_alias_ref()?);
-            }
-            Some(compact_type)
-        }
-        _ => Some(compact_type),
-    }
 }
